--- app/src/main/java/com/ably/tracking/demo/subscriber/common/StringExtensions.kt
@@ -1,4 +1,0 @@
-package com.ably.tracking.demo.subscriber.common
-
-fun String.canParseToDouble() =
-    toDoubleOrNull() != null
