+++ app/src/main/java/com/ably/tracking/demo/subscriber/data/ably/AssetTrackerAnimator.kt
@@ -1,4 +1,4 @@
-package com.ably.tracking.demo.subscriber.ably
+package com.ably.tracking.demo.subscriber.data.ably
 
 import com.ably.tracking.LocationUpdate
 import com.ably.tracking.ui.animation.LocationAnimator
