+++ app/src/main/java/com/ably/tracking/demo/subscriber/data/ably/AssetTrackerAnimatorPosition.kt
@@ -1,4 +1,4 @@
-package com.ably.tracking.demo.subscriber.ably
+package com.ably.tracking.demo.subscriber.data.ably
 
 import com.ably.tracking.ui.animation.Position
