+++ app/src/main/java/com/ably/tracking/demo/subscriber/data/api/CreateOrderBody.kt
@@ -1,4 +1,4 @@
-package com.ably.tracking.demo.subscriber.api
+package com.ably.tracking.demo.subscriber.data.api
 
 import com.ably.tracking.demo.subscriber.domain.GeoCoordinates
