+++ app/src/main/java/com/ably/tracking/demo/subscriber/data/api/CreateOrderResponse.kt
@@ -1,4 +1,4 @@
-package com.ably.tracking.demo.subscriber.api
+package com.ably.tracking.demo.subscriber.data.api
 
 data class CreateOrderResponse(
     val orderId: Long,
