+++ app/src/main/java/com/ably/tracking/demo/subscriber/data/api/TokenResponse.kt
@@ -1,4 +1,4 @@
-package com.ably.tracking.demo.subscriber.api
+package com.ably.tracking.demo.subscriber.data.api
 
 data class TokenResponse(
     val token: String
