+++ app/src/main/java/com/ably/tracking/demo/subscriber/domain/OrderManager.kt
@@ -1,8 +1,8 @@
 package com.ably.tracking.demo.subscriber.domain
 
 import com.ably.tracking.TrackableState
-import com.ably.tracking.demo.subscriber.ably.AssetTracker
-import com.ably.tracking.demo.subscriber.api.DeliveryServiceDataSource
+import com.ably.tracking.demo.subscriber.data.ably.AssetTracker
+import com.ably.tracking.demo.subscriber.data.api.DeliveryServiceDataSource
 import com.ably.tracking.demo.subscriber.secrets.SecretsManager
 import kotlinx.coroutines.flow.map
