+++ app/src/main/java/com/ably/tracking/demo/subscriber/presentation/screen/dashboard/DashboardViewModel.kt
@@ -6,11 +6,10 @@ import androidx.lifecycle.viewModelScope
 import com.ably.tracking.ConnectionException
 import com.ably.tracking.LocationUpdate
 import com.ably.tracking.Resolution
-import com.ably.tracking.demo.subscriber.ably.AssetTrackerAnimator
-import com.ably.tracking.demo.subscriber.ably.AssetTrackerAnimatorPosition
-import com.ably.tracking.demo.subscriber.common.FixedSizeMutableList
-import com.ably.tracking.demo.subscriber.common.FusedLocationSource
 import com.ably.tracking.demo.subscriber.common.distanceTo
+import com.ably.tracking.demo.subscriber.data.FusedLocationSource
+import com.ably.tracking.demo.subscriber.data.ably.AssetTrackerAnimator
+import com.ably.tracking.demo.subscriber.data.ably.AssetTrackerAnimatorPosition
 import com.ably.tracking.demo.subscriber.domain.OrderManager
 import com.ably.tracking.demo.subscriber.domain.OrderState
 import com.ably.tracking.demo.subscriber.presentation.navigation.Navigator
