+++ app/src/test/java/com/ably/tracking/demo/subscriber/data/api/ApiDeliveryServiceDataSourceTest.kt
@@ -1,4 +1,4 @@
-package com.ably.tracking.demo.subscriber.api
+package com.ably.tracking.demo.subscriber.data.api
 
 import com.ably.tracking.demo.subscriber.BuildConfig
 import com.ably.tracking.demo.subscriber.domain.GeoCoordinates
