+++ app/src/test/java/com/ably/tracking/demo/subscriber/secrets/InMemorySecretsManagerTest.kt
@@ -1,6 +1,6 @@
 package com.ably.tracking.demo.subscriber.secrets
 
-import com.ably.tracking.demo.subscriber.api.FakeDeliveryServiceDataSource
+import com.ably.tracking.demo.subscriber.data.api.FakeDeliveryServiceDataSource
 import com.google.common.truth.Truth.assertThat
 import kotlinx.coroutines.ExperimentalCoroutinesApi
 import kotlinx.coroutines.test.runTest
